Unified env_helpers.ts for our typescript stuff
