+++ src/js/graphql_endpoint/modules/jwt.ts
@@ -1,6 +1,7 @@
 import * as express from "express";
 import * as jwt from "jsonwebtoken";
 import * as AWS from "aws-sdk";
+import { getAwsClient } from "./env_helpers";
 
 const IS_LOCAL = process.env.IS_LOCAL == "True" || null; // get this from environment
 const JWT_SECRET_ID = process.env.JWT_SECRET_ID; // get this from environment
@@ -9,17 +10,7 @@ const JWT_SECRET_ID = process.env.JWT_SECRET_ID; // get this from environment
 let JWT_SECRET: string = "";
 
 function getSecretsManagerClient(): AWS.SecretsManager {
-    // TODO: This code should be unified with the dynamodb stuff in schema_client;
-    // or perhaps even create a typescript equivalent of env_helpers
-    return new AWS.SecretsManager({
-        apiVersion: "2017-10-17",
-        region: IS_LOCAL ? process.env.AWS_REGION : undefined,
-        accessKeyId: IS_LOCAL ? process.env.GRAPL_AWS_ACCESS_KEY_ID : undefined,
-        secretAccessKey: IS_LOCAL
-            ? process.env.GRAPL_AWS_ACCESS_KEY_SECRET
-            : undefined,
-        endpoint: IS_LOCAL ? process.env.GRAPL_AWS_ENDPOINT : undefined,
-    });
+    return getAwsClient(AWS.SecretsManager);
 }
 
 async function fetchJwtSecret(): Promise<string> {
